+++ src/containers/StockList/StockList.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from "react";
 import styles from "../../components/stock/StockListLayout.module.css";
 import { symbols } from "../../components/stock/symbolList";
-import { FMP_KEY } from "../../keys/fmpkey";
+import { FMP_CLOUD_API_KEY } from "../apis/index";
 import StockListLayout from "../../components/stock/StockListLayout";
 import PropagateLoader from "react-spinners/PropagateLoader";
 import { Box } from "@material-ui/core";
 
-const url = `https://fmpcloud.io/api/v3/quote/${symbols},?apikey=${FMP_KEY}`;
+const url = `https://fmpcloud.io/api/v3/quote/${symbols},?apikey=${FMP_CLOUD_API_KEY}`;
 
 const StockLists = () => {
   const [stocks, setStocks] = useState([]);
@@ -44,9 +44,9 @@ const StockLists = () => {
           />
         </Box>
       ) : (
-        stocks.map((stock, index) => (
+        stocks.map((stock) => (
           <StockListLayout
-            key={index.toString()}
+            key={stock.symbol.toString()}
             symbol={stock.symbol}
             name={stock.name}
             price={stock.price}
