+++ src/containers/apis/index.js
@@ -1,0 +1,2 @@
+export const RAPID_BING_API_KEY = process.env.REACT_APP_RAPID_APIKEY;
+export const FMP_CLOUD_API_KEY = process.env.REACT_APP_FMP_KEY;
